Former-commit-id: 2071b86a97aab546d827553f81ac3013c0a791e6

+++ backend/src/main/java/com/bigdata/tikitacar/car/controller/DealController.java
@@ -2,18 +2,21 @@ package com.bigdata.tikitacar.car.controller;
 
 import com.bigdata.tikitacar.car.dto.request.DealRegisterRequestDto;
 import com.bigdata.tikitacar.car.dto.request.DealUpdateRequestDto;
+import com.bigdata.tikitacar.car.dto.request.DealUpdateStatusRequestDto;
 import com.bigdata.tikitacar.car.dto.response.DealSearchResponseDto;
 import com.bigdata.tikitacar.car.service.DealService;
-import com.bigdata.tikitacar.user.entity.User;
+import com.bigdata.tikitacar.user.dto.response.UserFindResponseDto;
 import com.bigdata.tikitacar.user.service.UserService;
 import com.bigdata.tikitacar.util.JwtService;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 @RestController
@@ -38,11 +41,23 @@ public class DealController {
         ResponseEntity response = null;
         Map<String,Object> map = new HashMap<String, Object>();
 
-        //유저정보 dto에 등록
-        dealRegisterRequestDto.updateSellerId(userService.findUserByEmail(jwtService.getEmailFromToken(token)).getId());
-        dealService.registerDeal(dealRegisterRequestDto);
+        String loginEamil = jwtService.getEmailFromToken(token);
 
-        response = new ResponseEntity(map, HttpStatus.OK);
+        if(loginEamil!=null){
+            //유저정보 dto에 등록
+            dealRegisterRequestDto.updateSellerId(userService.findUserByEmail(loginEamil).getId());
+            dealService.registerDeal(dealRegisterRequestDto);
+
+            map.put("msg","거래 등록에 성공했습니다.");
+            map.put("status","success");
+            response = new ResponseEntity(map, HttpStatus.OK);
+        }else{
+            map.put("msg","거래 등록에 실패했습니다.");
+            map.put("status","fail");
+            response = new ResponseEntity(map, HttpStatus.BAD_REQUEST);
+        }
+
+
         return response;
     }
 
@@ -56,6 +71,8 @@ public class DealController {
 
         DealSearchResponseDto dealSearchResponseDto = dealService.searchDeal(id);
 
+        map.put("msg","거래 조회에 성공했습니다.");
+        map.put("status","success");
         map.put("data",dealSearchResponseDto);
         response = new ResponseEntity(map,HttpStatus.OK);
 
@@ -65,25 +82,91 @@ public class DealController {
 
     //Update
     @ApiOperation("거래 업데이트")
-    @PutMapping("/")
-    public Object dealUpdate(@RequestBody DealUpdateRequestDto dealUpdateRequestDto){
+    @PutMapping("")
+    public Object dealUpdate(@RequestHeader(value="Authorization") String token,
+                             @RequestBody DealUpdateRequestDto dealUpdateRequestDto){
         ResponseEntity response = null;
         Map<String,Object> map = new HashMap<String, Object>();
 
-        dealService.updateDeal(dealUpdateRequestDto);
-        response = new ResponseEntity(map,HttpStatus.OK);
+        String loginEmail = jwtService.getEmailFromToken(token);
+        DealSearchResponseDto dealSearchResponseDto = dealService.searchDeal(dealUpdateRequestDto.getId());
+
+        if(loginEmail!= null && loginEmail.equals(dealSearchResponseDto.getEmail())){
+            dealService.updateDeal(dealUpdateRequestDto);
+            map.put("msg","거래 수정에 성공했습니다.");
+            map.put("status","success");
+            response = new ResponseEntity(map,HttpStatus.OK);
+        }else{
+            map.put("msg","거래 수정에 실패하였습니다.");
+            map.put("status","fail");
+            response = new ResponseEntity(map,HttpStatus.BAD_REQUEST);
+        }
 
         return response;
     }
 
-    //Delete
-    @ApiOperation("거래 삭제")
-    @DeleteMapping("/{id}")
-    public Object dealDelete(@PathVariable Long id){
+    @ApiOperation("거래 완료처리")
+    @PutMapping("/status/{id}")
+    public Object dealDone(@RequestHeader(value="Authorization") String token,
+                           @PathVariable("id") Long dealId,
+                           @RequestBody DealUpdateStatusRequestDto dealUpdateStatusRequestDto){
         ResponseEntity response = null;
         Map<String,Object> map = new HashMap<String, Object>();
 
-        dealService.removeDeal(id);
+        //현재 로그인한 아이디
+        String loginEmail = jwtService.getEmailFromToken(token);
+        DealSearchResponseDto dealSearchResponseDto = dealService.searchDeal(dealId);
+
+        if(loginEmail!=null && loginEmail.equals(dealSearchResponseDto.getEmail())){
+            dealService.updateDealStatus(dealId,dealUpdateStatusRequestDto);
+            map.put("msg","거래 완료에 성공했습니다.");
+            map.put("status","success");
+            response = new ResponseEntity(map,HttpStatus.OK);
+        }else{
+            map.put("msg","거래 완료에 실패했습니다.");
+            map.put("status","fail");
+            response = new ResponseEntity(map,HttpStatus.BAD_REQUEST);
+        }
+        return response;
+    }
+
+
+    //Delete
+    @ApiOperation("거래 삭제")
+    @DeleteMapping("/{id}")
+    public Object dealDelete(@RequestHeader(value="Authorization") String token,
+                             @PathVariable Long id){
+        ResponseEntity response = null;
+        Map<String,Object> map = new HashMap<String, Object>();
+
+        String loginEmail = jwtService.getEmailFromToken(token);
+        DealSearchResponseDto dealSearchResponseDto = dealService.searchDeal(id);
+
+        if(loginEmail!= null && loginEmail.equals(dealSearchResponseDto.getEmail())){
+            dealService.removeDeal(id);
+            map.put("msg","거래 삭제에 성공했습니다.");
+            map.put("status","success");
+            response = new ResponseEntity(map,HttpStatus.OK);
+
+        }else{
+            map.put("msg","거래 삭제에 실패했습니다.");
+            map.put("status","fail");
+            response = new ResponseEntity(map,HttpStatus.BAD_REQUEST);
+        }
+
+        return response;
+    }
+
+    //GetAll
+    @ApiOperation("거래 리스트 조회")
+    @GetMapping("/all/{page}")
+    public Object dealGetAll(@PathVariable("page") int page){
+        ResponseEntity response = null;
+        Map<String,Object> map = new HashMap<String, Object>();
+
+        List<DealSearchResponseDto> dealSearchResponseDtoList = dealService.searchAll(PageRequest.of(page, 10));
+
+        map.put("data",dealSearchResponseDtoList);
         response = new ResponseEntity(map,HttpStatus.OK);
 
         return response;
